+++ backend/runtimeconsumer/lock_unix.go
@@ -44,5 +44,5 @@ func acquireLock(directory string) (func(), error) {
 
 func ownedFile(info os.FileInfo) bool {
 	stat, ok := info.Sys().(*syscall.Stat_t)
-	return ok && stat.Uid == uint32(os.Getuid())
+	return ok && int64(stat.Uid) == int64(os.Getuid())
 }
